Add more hooks to tasks

+++ Launcher/src/Launcher/Tasks/Implementations/Launch/UpdateServersTask.cs
@@ -4,7 +4,9 @@ using System.IO;
 using System.Threading.Tasks;
 using HashedFiles;
 using IO;
+using KludgeBox.Events.Global;
 using Launcher.Nodes;
+using PatchApi.Events;
 
 #endregion
 
@@ -20,6 +22,14 @@ public class UpdateServersTask : LauncherTask
     {
         var serversUrl = Main.ApiProvider.GetServersUrl();
         var targetFile = Path.Combine(Paths.MinecraftDirPath.AsAbsolute(), "servers.dat");
+
+        var evt = new ServersUpdatingEvent(targetFile, serversUrl);
+        
+        if(EventBus.PublishIsCancelled(evt))
+            return;
+
+        targetFile = evt.LocalServersFilePath;
+        serversUrl = evt.RemoteServersFileDownloadUrl;
         
         if (File.Exists(targetFile))
             File.Delete(targetFile);
